refactor(frontend): extract chat stream logic into useChatStream hook
ChatInterfaceInner was 400 lines with message building, SSE parsing,
and RAG status handling inline. Move all send/stream logic to a
dedicated useChatStream hook, reducing the component to pure wiring.
